yolo_detect/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
flask>=2.3.0
pyyaml>=6.0
